requirements.txt
yfinance>=0.2.40
pandas>=2.1
numpy>=1.26
pyarrow>=15.0
pyyaml>=6.0
requests>=2.31
alpaca-py>=0.28
anthropic>=0.40
python-dateutil>=2.9
pytz>=2024.1
